SL 'interpretations' language
I've tried to nail down and make consistent my use of 'interpretation' language in SL. (After a conversation with Dave, I think it may be idiosyncratic, but I like it.) I want to think of interpretations as rich, maybe even as being QL interpretations, but where we just ignore everything but the valuation functions they provide. Then we can have a consistent definition of entailment in terms of interpretations, across SL and QL. I think this version is cleaner/more precise about this. Biggest changes are in ch. 4, but some in 2, 3, and 9 too.